+++ opennmt/tests/sequence_to_sequence_test.py
@@ -38,6 +38,24 @@ class SequenceToSequenceTest(tf.test.TestCase):
       self.assertListEqual(
           [b"Mon", b"name", b"is", b"Max", b".", b""], replaced_target_tokens[1].tolist())
 
+  def _testPharaohAlignments(self, line, lengths, expected_matrix):
+    matrix = sequence_to_sequence.alignment_matrix_from_pharaoh(
+        tf.constant(line), lengths[0], lengths[1])
+    matrix = tf.cast(matrix, tf.int32)
+    with self.test_session() as sess:
+      self.assertListEqual(expected_matrix, sess.run(matrix).tolist())
+
+  def testPharaohAlignments(self):
+    self._testPharaohAlignments("0-0", [1, 1], [[1]])
+    self._testPharaohAlignments(
+        "0-0 1-1 2-2 3-3", [4, 4], [[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 1, 0], [0, 0, 0, 1]])
+    self._testPharaohAlignments(
+        "0-0 1-1 2-3 3-2", [4, 4], [[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 0, 1], [0, 0, 1, 0]])
+    self._testPharaohAlignments(
+        "0-0 1-2 1-1", [2, 3], [[1, 0], [0, 1], [0, 1]])
+    self._testPharaohAlignments(
+        "0-0 1-2 1-1 2-4", [3, 5], [[1, 0, 0], [0, 1, 0], [0, 1, 0], [0, 0, 0], [0, 0, 1]])
+
 
 if __name__ == "__main__":
   tf.test.main()